SnowyDune 21/11

+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/service/HotelService.java
@@ -25,7 +25,7 @@ public class HotelService {
 		return hotelRepository.listHotelsStatus(status);
 	}
 	
-	public void updateHotelStatus(Long id ,Long status) {
+	public void updateHotelStatus(Long id ,Integer status) {
 		hotelRepository.updateHotelStatus(id, status);
 	}
